Change the design of the project by create new package folder with the same name of the main project name and create __init__ file inside with many changed all explain on info.md file.
