+++ flaskinfo/accounts/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from pro import db
+from flaskinfo import db
 
 
 class User(db.Model):
@@ -12,7 +12,7 @@ class User(db.Model):
     email = db.Column(db.String(80), unique=True, nullable=False)
     image = db.Column(db.String(22), nullable=False, default='default.jpg')
     password = db.Column(db.String(40), nullable=False)
-    post = db.relastionship('Post', backref='author', lazy=True)
+    posts = db.relationship('Post', backref='author', lazy=True)
 
     def __repr__(self):
         return f"User('{self.username}', '{self.email}', '{self.image}')"
@@ -28,7 +28,7 @@ class Post(db.Model):
     content = db.Column(db.Text, nullable=False)
     photo = db.Column(db.String(42), nullable=False, default='default.jpg')
     date = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    user_id = db.Column(db.Integer, db.Foreignkey('user.id'), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
 
     def __repr__(self):
         return f"User('{self.title}', '{self.date}', '{self.photo}')"
