+++ flaskinfo/static/css/album.css
