+++ flaskinfo/templates/auth/login_form.html
