+++ flaskinfo/templates/auth/register_form.html
@@ -14,7 +14,7 @@
 				<form method="POST" action="">
 					{{ form.hidden_tag() }}
 					<fieldset class='form-group'>
-						<legend class="border-bottom mb-4">Login</legend>
+						<legend class="border-bottom mb-4">Registration</legend>
 						<div class="form-group">
 							{{ form.username.label(class="form-control-label") }}
