+++ flaskinfo/templates/base.html
