+++ flaskinfo/templates/footer.html
