+++ flaskinfo/templates/index.html
