+++ flaskinfo/templates/navbar.html
